Contrived integer overflows (backport to 0.5)

+++ tempesta_fw/classifier/frang.c
@@ -1226,9 +1226,11 @@ frang_start(void)
 {
 	/* Convert these timeouts to jiffies for convenience */
 	frang_cfg.clnt_hdr_timeout =
-		*(unsigned int *)&frang_cfg.clnt_hdr_timeout * HZ;
+		*(unsigned int *)&frang_cfg.clnt_hdr_timeout *
+		(unsigned long)HZ;
 	frang_cfg.clnt_body_timeout =
-		*(unsigned int *)&frang_cfg.clnt_body_timeout * HZ;
+		*(unsigned int *)&frang_cfg.clnt_body_timeout *
+		(unsigned long)HZ;
 	return 0;
 }
